move endpoints under api/
