+++ education-ai-suite/smart-classroom/components/board_ocr/summary_with_ocr.py
@@ -1,34 +1,7 @@
-from typing import Optional
-
-from fastapi import APIRouter, Header
-from fastapi.responses import JSONResponse
-
 from components.board_ocr.feature_base import Capability, FeatureModule
-from components.board_ocr.board_ocr_service import read_board_ocr, summarize_board_ocr
+from api.board_ocr import board_ocr_router
 
-board_ocr_router = APIRouter()
-
-
-@board_ocr_router.get("/board-ocr/ocr")
-def get_board_ocr_endpoint(
-    x_session_id: Optional[str] = Header(None),
-):
-    # Task 1 — return the board (content-screen) OCR extraction + status.
-    return JSONResponse(
-        content=read_board_ocr(x_session_id), status_code=200
-    )
-
-
-@board_ocr_router.post("/board-ocr/summary")
-def board_ocr_summary_endpoint(
-    x_session_id: Optional[str] = Header(None),
-):
-    # Task 2 — summarize the board OCR text via VLM/LLM (placeholder).
-    return JSONResponse(
-        content=summarize_board_ocr(x_session_id), status_code=200
-    )
-
-
+"""Summary with OCR feature module. Placeholder for future feature based design"""
 class SummaryWithOcrFeature(FeatureModule):
     id = "summary_with_ocr"
     requires = [Capability.OCR, Capability.TEXT_GEN]
